socialdistancing/requirements.txt
pandas >= 1.1.0
numpy >= 1.19.1
Flask == 1.1.2
google-cloud-storage == 1.33.0